requirements.txt
numpy ~= 1.14.3
scipy ~= 1.2.0
future ~= 0.16.0
matplotlib ~= 2.2.2
pillow
opencv-python ~= 3.4.0
scikit-image ~= 0.14.0
pyaml
pytorch==1.9.0
torchvision==0.10.0
tensorboardX
